requirements.txt
streamlit>=1.31.0
requests
beautifulsoup4
pillow
streamlit-pdf-viewer
